Make apache-2.0 and lgpl rules more specific
These rules are AND rules. Thus they should only match if both
licenses are actually detected in the text.

There are a bunch of rules that have similar problems. Since the rules
touched in this diff caused problems for us,
I am just touching them and and consider remaining rules out of scope.

Signed-off-by: Alexander zur Bonsen <alexander.zur.bonsen@tngtech.com>